requirements/base.txt
# Core
pytz==2020.1
Django==3.0.8
Pillow==7.2.0

# Configuration manager 
django-configurations==2.2
python-decouple==3.3

# For the persistence stores
psycopg2-binary==2.8.5
dj-database-url==0.5.0

# Model Tools
django-autoslug==1.9.7
django-model-utils==4.0.0

# Rest apis
djangorestframework==3.11.0
Markdown==3.1.1
django-filter==2.2.0
django-cors-headers==3.4.0
django-extra-fields==2.0.5
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1

